requirements.txt
streamlit==1.32.0
python-docx==1.1.0
PyPDF2==3.0.1
fb2reader==1.0.3
scikit-learn==1.4.1.post1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
requests==2.31.0 